requirements.txt
pandas
requests
apache-airflow

